+++ bin-proto/src/types/mod.rs
@@ -10,6 +10,4 @@ mod option;
 mod smart_ptr;
 mod string;
 mod tuple;
-#[cfg(feature = "uuid")]
-mod uuid;
 mod vec;