+++ tests/src/lib.rs
@@ -5,6 +5,8 @@ mod ctx;
 #[cfg(test)]
 mod enum_trait;
 #[cfg(test)]
+mod enums;
+#[cfg(test)]
 mod flexible_array_member;
 #[cfg(test)]
 mod ipv4;